refactor(plugins): isolate authoritative host-link repair owners
