+++ src/commands/doctor-plugin-registry.ts
@@ -10,22 +10,11 @@ import { saveJsonFile } from "../infra/json-file.js";
 import { tryReadJsonSync } from "../infra/json-files.js";
 import type { BundledPluginSource } from "../plugins/bundled-sources.js";
 import {
-  resolveDefaultPluginExtensionsDir,
-  resolveDefaultPluginNpmDir,
-} from "../plugins/install-paths.js";
-import {
   loadInstalledPluginIndexInstallRecords,
   type InstalledPluginIndexRecordStoreOptions,
 } from "../plugins/installed-plugin-index-records.js";
 import { loadInstalledPluginIndex } from "../plugins/installed-plugin-index.js";
 import { hasRetainedManagedNpmInstallMarker } from "../plugins/managed-npm-retention.js";
-import { listManagedPluginNpmRootsSync } from "../plugins/npm-project-roots.js";
-import {
-  auditOpenClawPeerDependenciesInManagedNpmRoot,
-  type OpenClawPeerLinkAuditIssue,
-  reconcileRegisteredOpenClawHostLinks,
-  relinkOpenClawPeerDependenciesInManagedNpmRoot,
-} from "../plugins/plugin-peer-link.js";
 import { refreshPluginRegistry } from "../plugins/plugin-registry.js";
 import {
   listStaleLocalBundledPluginInstallRecords,
@@ -40,6 +29,11 @@ import {
   staleManagedNpmInstallGenerationToRepairEffect,
   type StaleManagedNpmInstallGenerationIssue,
 } from "./doctor-plugin-generations.js";
+import {
+  listManagedPluginNpmRoots,
+  listPluginOpenClawHostLinkIssues,
+  maybeRepairPluginOpenClawHostLinks,
+} from "./doctor-plugin-host-links.js";
 import type { DoctorPrompter } from "./doctor-prompter.js";
 import {
   migratePluginRegistryForInstall,
@@ -59,11 +53,6 @@ type StaleManagedNpmBundledPlugin = {
   packageDir: string;
   npmRoot: string;
   version?: string;
-};
-
-type PluginRegistryDoctorNoteLogger = {
-  info: (message: string) => void;
-  warn: (message: string) => void;
 };
 
 type PluginRegistryHealthIssue =
@@ -108,14 +97,6 @@ type PluginRegistryHealthIssue =
     }
   | StaleManagedNpmInstallGenerationIssue;
 
-type ManagedNpmPackageReadFailure = {
-  packageDir: string;
-  reason: string;
-};
-
-const formatPackageReadFailure = (error: unknown): string =>
-  error instanceof Error ? error.message : String(error);
-
 function readJsonObject(filePath: string): Record<string, unknown> | null {
   const parsed = tryReadJsonSync(filePath);
   return isRecord(parsed) ? parsed : null;
@@ -132,22 +113,6 @@ function readStringMap(value: unknown): Record<string, string> {
     }
   }
   return result;
-}
-
-function resolveManagedPluginNpmRoot(params: PluginRegistryDoctorRepairParams): string {
-  return params.stateDir
-    ? path.join(params.stateDir, "npm")
-    : resolveDefaultPluginNpmDir(params.env);
-}
-
-function resolveRegisteredPluginExtensionsRoot(params: PluginRegistryDoctorRepairParams): string {
-  return params.stateDir
-    ? path.join(params.stateDir, "extensions")
-    : resolveDefaultPluginExtensionsDir(params.env);
-}
-
-function listManagedPluginNpmRoots(params: PluginRegistryDoctorRepairParams): string[] {
-  return listManagedPluginNpmRootsSync(resolveManagedPluginNpmRoot(params));
 }
 
 function deleteObjectKey(record: Record<string, unknown>, key: string): boolean {
@@ -391,117 +356,6 @@ async function maybeRepairStaleLocalBundledPluginInstallRecords(
   return stale.map((record) => record.pluginId);
 }
 
-/** Relinks npm-owned plugin packages to the current OpenClaw host package. */
-export async function maybeRepairPluginOpenClawHostLinks(
-  params: PluginRegistryDoctorRepairParams,
-): Promise<boolean> {
-  const npmRoots = listManagedPluginNpmRoots(params);
-  if (!params.prompter.shouldRepair) {
-    const audit = await listPluginOpenClawHostLinkIssues(params);
-    if (audit.peerLinkIssues.length > 0) {
-      note(
-        [
-          "Managed npm OpenClaw host peer links need repair:",
-          ...audit.peerLinkIssues.map((issue) => `- ${issue.packageName}: ${issue.reason}`),
-          `Repair with ${formatCliCommand("openclaw doctor --fix")} to relink managed npm plugin packages.`,
-        ].join("\n"),
-        "Plugin registry",
-      );
-    }
-    if (audit.packageReadFailures.length > 0) {
-      note(
-        [
-          "Managed npm plugin packages could not be inspected:",
-          ...audit.packageReadFailures.map(
-            (failure) => `- ${shortenHomePath(failure.packageDir)}: ${failure.reason}`,
-          ),
-        ].join("\n"),
-        "Plugin registry",
-      );
-    }
-    if (audit.registeredPackageReadFailures.length > 0) {
-      note(
-        [
-          "Registered npm plugin packages could not be inspected:",
-          ...audit.registeredPackageReadFailures.map(
-            (failure) => `- ${shortenHomePath(failure.packageDir)}: ${failure.reason}`,
-          ),
-        ].join("\n"),
-        "Plugin registry",
-      );
-    }
-    if (audit.registeredPeerLinkIssues.length > 0) {
-      note(
-        [
-          "Registered npm plugin OpenClaw host links need repair:",
-          ...audit.registeredPeerLinkIssues.map(
-            (issue) => `- ${issue.packageName}: ${issue.reason}`,
-          ),
-          `Repair with ${formatCliCommand("openclaw doctor --fix")} to relink registered npm plugin packages.`,
-        ].join("\n"),
-        "Plugin registry",
-      );
-    }
-    return false;
-  }
-
-  const messages: { level: "info" | "warn"; message: string }[] = [];
-  const logger: PluginRegistryDoctorNoteLogger = {
-    info: (message) => messages.push({ level: "info", message }),
-    warn: (message) => messages.push({ level: "warn", message }),
-  };
-  const results = await Promise.all(
-    npmRoots.map((npmRoot) =>
-      relinkOpenClawPeerDependenciesInManagedNpmRoot({
-        npmRoot,
-        logger,
-        onPackageReadError: (error, packageDir) => {
-          logger.warn(
-            `Could not inspect managed npm package ${shortenHomePath(packageDir)}: ${formatPackageReadFailure(error)}`,
-          );
-        },
-      }),
-    ),
-  );
-  const repaired = results.reduce((total, result) => total + result.repaired, 0);
-  const registeredRepair = await reconcileRegisteredOpenClawHostLinks({
-    installRecords: await loadInstalledPluginIndexInstallRecords(params),
-    extensionsDir: resolveRegisteredPluginExtensionsRoot(params),
-    env: params.env,
-    mode: "repair",
-    logger,
-    onPackageReadError: (error, packageDir) => {
-      logger.warn(
-        `Could not inspect registered npm package ${shortenHomePath(packageDir)}: ${formatPackageReadFailure(error)}`,
-      );
-    },
-  });
-
-  if (repaired > 0) {
-    note(
-      `Repaired OpenClaw host peer link(s) for ${repaired} managed npm plugin package(s).`,
-      "Plugin registry",
-    );
-  }
-  if (registeredRepair.repaired > 0) {
-    note(
-      `Repaired OpenClaw host peer link(s) for ${registeredRepair.repaired} registered npm plugin package(s).`,
-      "Plugin registry",
-    );
-  }
-  const warnings = messages
-    .filter((message) => message.level === "warn")
-    .map((message) => `- ${message.message}`);
-  if (warnings.length > 0) {
-    note(
-      ["Could not repair all managed npm OpenClaw host peer links:", ...warnings].join("\n"),
-      "Plugin registry",
-    );
-  }
-
-  return repaired > 0 || registeredRepair.repaired > 0;
-}
-
 async function loadInstallRecordsWithoutPluginIds(
   params: PluginRegistryDoctorRepairParams,
   pluginIds: readonly string[],
@@ -511,47 +365,6 @@ async function loadInstallRecordsWithoutPluginIds(
     delete records[pluginId];
   }
   return records;
-}
-
-async function listPluginOpenClawHostLinkIssues(params: PluginRegistryDoctorRepairParams): Promise<{
-  peerLinkIssues: OpenClawPeerLinkAuditIssue[];
-  packageReadFailures: ManagedNpmPackageReadFailure[];
-  registeredPeerLinkIssues: OpenClawPeerLinkAuditIssue[];
-  registeredPackageReadFailures: ManagedNpmPackageReadFailure[];
-}> {
-  const packageReadFailures: ManagedNpmPackageReadFailure[] = [];
-  const registeredPackageReadFailures: ManagedNpmPackageReadFailure[] = [];
-  const audits = await Promise.all(
-    listManagedPluginNpmRoots(params).map((npmRoot) =>
-      auditOpenClawPeerDependenciesInManagedNpmRoot({
-        npmRoot,
-        onPackageReadError: (error, packageDir) => {
-          packageReadFailures.push({
-            packageDir,
-            reason: formatPackageReadFailure(error),
-          });
-        },
-      }),
-    ),
-  );
-  const registeredAudit = await reconcileRegisteredOpenClawHostLinks({
-    installRecords: await loadInstalledPluginIndexInstallRecords(params),
-    extensionsDir: resolveRegisteredPluginExtensionsRoot(params),
-    env: params.env,
-    mode: "audit",
-    onPackageReadError: (error, packageDir) => {
-      registeredPackageReadFailures.push({
-        packageDir,
-        reason: formatPackageReadFailure(error),
-      });
-    },
-  });
-  return {
-    peerLinkIssues: audits.flatMap((audit) => audit.issues),
-    packageReadFailures,
-    registeredPeerLinkIssues: registeredAudit.issues,
-    registeredPackageReadFailures,
-  };
 }
 
 export async function detectPluginRegistryHealthIssues(
